feat: move 5 FeedFilter subclasses to commons (batch 4)
Move the following filters from amethyst to commons module:
- BookmarkPrivateFeedFilter
- BookmarkPublicFeedFilter
- OldBookmarkPrivateFeedFilter
- OldBookmarkPublicFeedFilter
- SpammerAccountsFeedFilter

IAccount additions:
- bookmarkState: BookmarkListState
- oldBookmarkState: OldBookmarkListState
- transientHiddenUsers: Set<String>
- cacheProvider: ICacheProvider

Original locations retain typealiases for backwards compatibility.
Part of the KMP iOS migration effort.
